Models, dao's and database classes has been added

+++ app/src/main/java/com/example/openmarket/data/Comment.kt
@@ -8,9 +8,8 @@ import java.io.Serializable
 
 @Entity(tableName = "comments")
 data class Comment (
-        @PrimaryKey(autoGenerate = true)  @ColumnInfo(name = "id")  val id:Int,
+        @PrimaryKey(autoGenerate = true)  @ColumnInfo(name = "id")  val id:Long,
         @ColumnInfo(name = "comment_data") val commentdata:String,
         @ColumnInfo(name ="comment_date" )val dateOfComment:String,
         @ColumnInfo(name = "user_name") val userName:String
-
 ):Serializable
